+++ src/js_realm_object.hpp
@@ -110,7 +110,6 @@ struct RealmObjectClass : ClassDefinition<T, realm::js::RealmObject<T>> {
         {"linkingObjectsCount", wrap<linking_objects_count>},
         {"_isSameObject", wrap<is_same_object>},
         {"_objectKey", wrap<get_object_key>},
-        {"_objectId", wrap<get_object_id>},
         {"_setLink", wrap<set_link>},
         {"addListener", wrap<add_listener>},
         {"removeListener", wrap<remove_listener>},
@@ -351,22 +350,6 @@ void RealmObjectClass<T>::get_object_key(ContextType ctx, ObjectType object, Arg
     const Obj& obj = realm_object->obj();
     auto obj_key = obj.get_key();
     return_value.set(std::to_string(obj_key.value));
-}
-
-template <typename T>
-void RealmObjectClass<T>::get_object_id(ContextType ctx, ObjectType object, Arguments& args,
-                                        ReturnValue& return_value)
-{
-    args.validate_maximum(0);
-
-    auto realm_object = get_internal<T, RealmObjectClass<T>>(ctx, object);
-    if (!realm_object) {
-        throw std::runtime_error("Invalid 'this' object");
-    }
-
-    const Obj& obj = realm_object->obj();
-    auto obj_id = obj.get_object_id();
-    return_value.set(obj_id.to_string());
 }
 
 template <typename T>